We use now the blueprints framework
