+++ lmfdb/siegel_modular_forms/templates/Gamma0_2_avail_forms.html
