+++ lmfdb/siegel_modular_forms/templates/Gamma0_2_basic.html
