+++ lmfdb/siegel_modular_forms/templates/Gamma0_3_psi_3_avail_forms.html
