+++ lmfdb/siegel_modular_forms/templates/Gamma0_4_basic.html
