+++ lmfdb/siegel_modular_forms/templates/Gamma0_4_psi_4_avail_forms.html
