+++ lmfdb/siegel_modular_forms/templates/Gamma_2_basic.html
