+++ lmfdb/siegel_modular_forms/templates/ModularForm_GSp4_Q_basic.html
@@ -1,7 +1,7 @@
 {% extends 'homepage.html' %}
 {% block content %}
  
-{% include ['ModularForm_GSp4_Q/', group, '_basic.html']|join %}
+{% include [ group, '_basic.html']|join %}
   
 {% endblock %}
