+++ lmfdb/siegel_modular_forms/templates/ModularForm_GSp4_Q_gen_rel.html
@@ -1,7 +1,7 @@
 {% extends 'homepage.html' %}
 {% block content %}
 
-{% include ['ModularForm_GSp4_Q/', group, '_gen_rel.html']|join %}
+{% include [ group, '_gen_rel.html']|join %}
 
 {% endblock %}
