+++ lmfdb/siegel_modular_forms/templates/ModularForm_GSp4_Q_specimen.html
@@ -123,7 +123,7 @@
 </div></div>
 
 {% else %}
-{% include 'ModularForm_GSp4_Q/None.html' %}
+{% include 'None.html' %}
 {% endif %}
 
 {% endblock %}
