+++ lmfdb/siegel_modular_forms/templates/README.TXT
