+++ lmfdb/siegel_modular_forms/templates/Sp4Z_2_avail_forms.html
