+++ lmfdb/siegel_modular_forms/templates/Sp4Z_avail_forms.html
