+++ lmfdb/siegel_modular_forms/templates/Sp6Z_avail_forms.html
