+++ lmfdb/siegel_modular_forms/templates/Sp8Z_avail_forms.html
