--- lmfdb/templates/ModularForm_GSp4_Q/ModularForm_GSp4_Q_forms.html
@@ -1,13 +1,0 @@
-{% extends 'homepage.html' %}
-{% block content %}
-
-{% if not error %}
-{% include ['ModularForm_GSp4_Q/', group, '_avail_forms.html']|join %}
-{% else %}
-{% include 'ModularForm_GSp4_Q/None.html' %}
-{% endif %}
-
-{% endblock %}
-
-
-
